requirements.txt
streamlit
openai
langdetect
gspread
google-auth